+++ .github/workflows/basic-checks.yml
@@ -3,7 +3,7 @@ name: Basic checks
 on: [pull_request]
 
 env:
-  JAVA_VERSION: 16
+  JAVA_VERSION: 17
 
 jobs:
   spotless:
