+++ jda5/src/main/java/io/github/yusufsdiscordbot/yusufsdiscordcore/bot/button/interaction/YusufButtonClickEvent.java
@@ -1,12 +1,10 @@
-package io.github.yusufsdiscordbot.yusufsdiscordcore.bot.slash_command.button.interaction;
+package io.github.yusufsdiscordbot.yusufsdiscordcore.bot.button.interaction;
 
-import io.github.yusufsdiscordbot.yusufsdiscordcore.bot.slash_command.button.YusufButton;
+import io.github.yusufsdiscordbot.yusufsdiscordcore.bot.button.YusufButton;
 import io.github.yusufsdiscordbot.yusufsdiscordcore.bot.slash_command.interactions.Command;
-import net.dv8tion.jda.api.entities.MessageEmbed;
 import net.dv8tion.jda.api.events.interaction.ButtonClickEvent;
 
-import javax.annotation.Nonnull;
-
+@SuppressWarnings({"unused"})
 public class YusufButtonClickEvent extends YusufButtonInteraction {
     private final Command command;
     private final ButtonClickEvent buttonClickEvent;
@@ -35,27 +33,5 @@ public class YusufButtonClickEvent extends YusufButtonInteraction {
 
     public YusufButton getYusufButton() {
         return new YusufButton(buttonClickEvent.getButton());
-    }
-
-    public void replyQueuedMessage(@Nonnull String message) {
-        this.buttonClickEvent.reply(message).queue();
-    }
-
-    /**
-     * replays as an ephemeral message.
-     */
-    public void replyQueuedEphemeral(@Nonnull String message) {
-        this.buttonClickEvent.reply(message).setEphemeral(true).queue();
-    }
-
-    /**
-     * replays as an embed message.
-     */
-    public void replyQueuedEmbed(@Nonnull MessageEmbed messageEmbed) {
-        this.buttonClickEvent.replyEmbeds(messageEmbed).queue();
-    }
-
-    public void replyQueuedEphemeralEmbed(@Nonnull MessageEmbed messageEmbed) {
-        this.buttonClickEvent.replyEmbeds(messageEmbed).setEphemeral(true).queue();
     }
 }
