+++ jda5/src/main/java/io/github/yusufsdiscordbot/yusufsdiscordcore/bot/select_menu/YusufSelectionMenuEvent.java
@@ -1,0 +1,5 @@
+package io.github.yusufsdiscordbot.yusufsdiscordcore.bot.select_menu;
+
+//TODO : make this class
+public class YusufSelectionMenuEvent {
+}
