+++ jda5/src/main/java/io/github/yusufsdiscordbot/yusufsdiscordcore/bot/select_menu/interaction/YusufSelectionMenuInteraction.java
@@ -1,0 +1,5 @@
+package io.github.yusufsdiscordbot.yusufsdiscordcore.bot.select_menu.interaction;
+
+//TODO: make this class
+public class YusufSelectionMenuInteraction {
+}
