+++ jda5/src/main/java/io/github/yusufsdiscordbot/yusufsdiscordcore/bot/slash_command/command_option/YusufOptionMapping.java
@@ -36,7 +36,7 @@ public record YusufOptionMapping(OptionMapping optionMapping) {
         return optionMapping.getName();
     }
 
-    //TODO: fix this issue
+    // TODO: fix this issue
     @Nullable
     public YusufMember getAsMember() {
         return new YusufMember(optionMapping.getAsMember());
