+++ jda5/src/main/java/io/github/yusufsdiscordbot/yusufsdiscordcore/bot/slash_command/example/ExampleCommand.java
@@ -13,12 +13,15 @@
 
 package io.github.yusufsdiscordbot.yusufsdiscordcore.bot.slash_command.example;
 
+import io.github.yusufsdiscordbot.yusufsdiscordcore.bot.YusufInteraction;
+import io.github.yusufsdiscordbot.yusufsdiscordcore.bot.button.interaction.YusufButtonClickEvent;
 import io.github.yusufsdiscordbot.yusufsdiscordcore.bot.slash_command.core.YusufUser;
 import io.github.yusufsdiscordbot.yusufsdiscordcore.bot.slash_command.interactions.Command;
 import io.github.yusufsdiscordbot.yusufsdiscordcore.bot.slash_command.interactions.YusufSlashCommandEvent;
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.interactions.commands.OptionType;
 import net.dv8tion.jda.api.interactions.commands.build.OptionData;
+import net.dv8tion.jda.api.requests.restaction.interactions.ReplyAction;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.HashMap;
@@ -34,7 +37,7 @@ class ExampleCommand extends Command {
 
         getCommandData().addOptions(
                 new OptionData(OptionType.STRING, EXAMPLE_OPTION, "This is an example option")
-                    .addChoice("Test", "Works"));
+                        .addChoice("Test", "Works"));
     }
 
     @Override
@@ -62,5 +65,10 @@ class ExampleCommand extends Command {
         }
 
         yusufSlashCommandEvent.replyQueuedEmbed(builder.build());
+    }
+
+    @Override
+    public void onButtonClick(YusufButtonClickEvent yusufButtonClickEvent) {
+        //example
     }
 }
